release v3
added a picture frame to help with the puzzle of colour withotu forcing the player to use clues,  the waypoint now has a visible marker as well as a light
there is now a timer. so people can compare their scores vs their time.
removed glitchy collision with final door and stairs